requirements.txt
google-generativeai>=0.4.0
pypdf>=4.0.0
rich>=13.0.0
requests>=2.31.0
PyMuPDF>=1.23.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.9
